requirements.txt
# IPLive - Dependencies (Offline Mode)

# Flask web framework
Flask==2.3.3
Werkzeug==2.3.7

# MySQL database connector
mysql-connector-python==8.1.0

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-mock==3.11.1

# Production server
gunicorn==21.2.0

# Note: pycricbuzz is NOT required - app uses MockCricbuzz for offline mode
